Debug mode works

+++ GAME3001-W2023-Assignment02/src/PlayScene.h
@@ -29,7 +29,7 @@ private:
 	// IMGUI Function
 	void GUI_Function();
 	std::string m_guiTitle;
-	glm::vec2 m_mousePosition;
+	glm::ivec2 m_mousePosition;
 
 	// Game Objects for the Scene
 	Target* m_pTarget;
